cpu.rs: don't build the arm module when on non-arm
For initial ppc64le support, we are assuming POWER8 and above, therefore
we do not (yet?) need a specific Features struct for the architecture.
However, this arm struct will complain that it is never constructed, and
there are unused functions. Therefore, we don't build it on non-arm.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.